requirements.txt
asgiref==3.7.2
bleach==6.1.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.7
Django==5.0.2
django-admin-logs==1.3.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-user-activity-tracker==1.0.1
drf_api_logger==1.1.15
et-xmlfile==1.1.0
Faker==24.2.0
filelock==3.13.1
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
platformdirs==4.1.0
plotly==5.22.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
retrying==1.3.4
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.3.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
webencodings==0.5.1
Werkzeug==3.0.3
zipp==3.18.2
